+++ pycharm/src/main/java/rw/session/Session.java
@@ -1,17 +1,23 @@
 package rw.session;
 
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.VisibleForTesting;
 import rw.audit.RwSentry;
 import rw.handler.runConf.BaseRunConfHandler;
+import rw.session.cmds.Cmd;
+import rw.session.events.Action;
+import rw.session.events.*;
 
 import java.io.*;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.net.SocketException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 import static java.util.Map.entry;
@@ -21,62 +27,37 @@ class RawEvent {
     public String VERSION;
 }
 
-public class Session extends Thread {
-    private static final Logger LOGGER = Logger.getInstance(Session.class);
+class Client extends Thread {
+    private static final Logger LOGGER = Logger.getInstance(Client.class);
 
-    private final Project project;
+    private Socket socket;
 
-    private ServerSocket serverSocket;
-    private Socket clientSocket;
     private PrintWriter out;
     private BufferedReader in;
-    private Integer port = null;
-    private final BaseRunConfHandler handler;
-    private Map<String, Class<? extends Event>> events;
-    private Map<String, String> eventVersions;
+    ServerSocket server;
 
-    public Session(Project project, BaseRunConfHandler handler) {
-        this.project = project;
-        this.handler = handler;
+    Session session;
 
-        this.events = Map.ofEntries(
-                entry(Handshake.ID, Handshake.class),
-                entry(ModuleUpdate.ID, ModuleUpdate.class),
-                entry(FrameError.ID, FrameError.class),
-                entry(LineProfile.ID, LineProfile.class),
-                entry(StackUpdate.ID, StackUpdate.class),
-                entry(UserError.ID, UserError.class),
-                entry(ClearErrors.ID, ClearErrors.class),
-                entry(LineProfileClear.ID, LineProfileClear.class)
-        );
+    Client(ServerSocket server, Session session) {
+        this.session = session;
+        this.server = server;
+    }
 
-        this.eventVersions = Map.ofEntries(
-                entry(Handshake.ID, Handshake.VERSION),
-                entry(ModuleUpdate.ID, ModuleUpdate.VERSION),
-                entry(FrameError.ID, FrameError.VERSION),
-                entry(LineProfile.ID, LineProfile.VERSION),
-                entry(UserError.ID, UserError.VERSION),
-                entry(StackUpdate.ID, StackUpdate.VERSION),
-                entry(ClearErrors.ID, ClearErrors.VERSION),
-                entry(LineProfileClear.ID, LineProfileClear.VERSION)
-        );
-
+    public boolean connect() {
         try {
-            this.serverSocket = new ServerSocket(0);
-            this.port = this.serverSocket.getLocalPort();
+            this.socket = this.server.accept();
+            LOGGER.info("Client connected to the Session");
         } catch (IOException ignored) {
+            return false;
         }
+        return true;
     }
 
     public void run() {
-        if (this.serverSocket == null) {
-            return;
-        }
-
         try {
-            this.clientSocket = this.serverSocket.accept();
-            this.out = new PrintWriter(this.clientSocket.getOutputStream(), true);
-            this.in = new BufferedReader(new InputStreamReader(this.clientSocket.getInputStream()));
+
+            this.out = new PrintWriter(this.socket.getOutputStream(), true);
+            this.in = new BufferedReader(new InputStreamReader(this.socket.getInputStream()));
             String inputLine;
             try {
                 while ((inputLine = this.in.readLine()) != null) {
@@ -92,19 +73,87 @@ public class Session extends Thread {
         }
     }
 
-    public int getPort() {
-        return this.port;
-    }
-
     private void ingestLine(String line) {
         try {
-            Event event = this.eventFactory(line);
+            Event event = this.session.eventFactory(line);
             if (event == null) {
                 return;
             }
             event.handle();
         } catch (RuntimeException e) {
             RwSentry.get().captureException(e);
+        }
+    }
+
+    void send(Cmd cmd) {
+        Gson gson = new Gson();
+        String payload = gson.toJson(cmd);
+        LOGGER.info(String.format("Sending cmd %s", cmd.getId()));
+        this.out.println(payload);
+    }
+
+    public void close() {
+        try {
+            if (this.in != null) {
+                this.in.close();
+            }
+            if (this.out != null) {
+                this.out.close();
+            }
+            if (this.socket != null) {
+                this.socket.close();
+            }
+        } catch (Exception e) {
+            RwSentry.get().captureException(e);
+        }
+    }
+}
+
+public class Session extends Thread {
+    private static final Logger LOGGER = Logger.getInstance(Session.class);
+
+    private final Project project;
+
+    private ServerSocket serverSocket;
+    private Integer port = null;
+    private final BaseRunConfHandler handler;
+    private Map<String, Class<? extends Event>> events;
+    private Map<String, String> eventVersions;
+    private List<Client> clients;
+
+    public Session(Project project, BaseRunConfHandler handler) {
+        this.project = project;
+        this.handler = handler;
+        this.clients = new ArrayList<>();
+
+        this.events = Map.ofEntries(
+                entry(Handshake.ID, Handshake.class),
+                entry(ModuleUpdate.ID, ModuleUpdate.class),
+                entry(FrameError.ID, FrameError.class),
+                entry(LineProfile.ID, LineProfile.class),
+                entry(StackUpdate.ID, StackUpdate.class),
+                entry(UserError.ID, UserError.class),
+                entry(ClearErrors.ID, ClearErrors.class),
+                entry(LineProfileClear.ID, LineProfileClear.class),
+                entry(WatchingFiles.ID, WatchingFiles.class)
+        );
+
+        this.eventVersions = Map.ofEntries(
+                entry(Handshake.ID, Handshake.VERSION),
+                entry(ModuleUpdate.ID, ModuleUpdate.VERSION),
+                entry(FrameError.ID, FrameError.VERSION),
+                entry(LineProfile.ID, LineProfile.VERSION),
+                entry(UserError.ID, UserError.VERSION),
+                entry(StackUpdate.ID, StackUpdate.VERSION),
+                entry(ClearErrors.ID, ClearErrors.VERSION),
+                entry(LineProfileClear.ID, LineProfileClear.VERSION),
+                entry(WatchingFiles.ID, WatchingFiles.VERSION)
+        );
+
+        try {
+            this.serverSocket = new ServerSocket(0);
+            this.port = this.serverSocket.getLocalPort();
+        } catch (IOException ignored) {
         }
     }
 
@@ -134,22 +183,42 @@ public class Session extends Thread {
         return ret;
     }
 
+    public void run() {
+        if (this.serverSocket == null) {
+            return;
+        }
+
+        while (true) {
+            Client client = new Client(this.serverSocket, this);
+            if (!client.connect()) {
+                return;
+            }
+            client.start();
+            this.clients.add(client);
+        }
+    }
+
+    public void send(Cmd cmd) {
+        for (Client c : this.clients) {
+            c.send(cmd);
+        }
+    }
+
+    public int getPort() {
+        return this.port;
+    }
+
     public void close() {
         try {
-            if (this.in != null) {
-                this.in.close();
-            }
-            if (this.out != null) {
-                this.out.close();
-            }
-            if (this.clientSocket != null) {
-                this.clientSocket.close();
-            }
-            if (this.clientSocket != null) {
+            if (this.serverSocket != null) {
                 this.serverSocket.close();
             }
         } catch (Exception e) {
             RwSentry.get().captureException(e);
+        }
+
+        for (Client c : this.clients) {
+            c.close();
         }
     }
 }