+++ pycharm/src/main/java/rw/session/cmds/Cmd.java
@@ -1,0 +1,5 @@
+package rw.session.cmds;
+
+abstract public class Cmd {
+    public abstract String getId();
+}