+++ pycharm/src/main/java/rw/session/events/FrameError.java
@@ -1,10 +1,9 @@
-package rw.session;
+package rw.session.events;
 
 import com.intellij.openapi.diagnostic.Logger;
 import rw.dialogs.DialogFactory;
-import rw.stack.Frame;
 
-public class FrameError extends UserError {
+public class FrameError extends FileError {
     private static final Logger LOGGER = Logger.getInstance(FrameError.class);
 
     public static final String ID = "FrameError";
@@ -12,7 +11,7 @@ public class FrameError extends UserError {
 
     @Override
     public void handle() {
-        LOGGER.info("Handling FrameError");
+        LOGGER.info("Handling UserError " + String.format("(%s, %d)", this.getLocalPath(), this.getLine()));
         super.handle();
         DialogFactory.get().showFirstFrameErrorDialog(this.handler.getProject());
     }