+++ pycharm/src/main/java/rw/session/events/LineProfileClear.java
@@ -1,10 +1,8 @@
-package rw.session;
+package rw.session.events;
 
 import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.openapi.util.Pair;
 
 import java.util.List;
-import java.util.Map;
 
 public class LineProfileClear extends FileEvent {
     private static final Logger LOGGER = Logger.getInstance(LineProfileClear.class);