+++ pycharm/src/test/java/rw/tests/ui/TestMisc.java
@@ -33,8 +33,6 @@ public class TestMisc extends PackageTestBase {
     void installsPackage(final RemoteRobot remoteRobot) throws Exception {
         Depot.run("env.get_remote().clean()");
         Package.run("p.rm_config");
-        // Package is downloaded on application start so this will break things
-        // Plugin.run("p.rm_package");
         Package.run("p.push_release");
         String version = Package.run("v.current_version")[0];
         Depot.run("p.publish_release");